feat: factor out metrics into hook
