+++ frontend/app/src/pages/main/v1/workflow-runs-v1/hooks/use-column-filters.tsx
@@ -4,7 +4,7 @@ import {
   ColumnFiltersState,
   Updater,
 } from '@tanstack/react-table';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
 
 type FilterParams = {
